organizing BiRB, added BiRB tutorial
