+++ pygsti/algorithms/randomcircuit.py
@@ -3128,3 +3128,315 @@ def create_random_germpower_mirror_circuits(pspec, absolute_compilation, depths,
         outlist.append(idealout)
 
     return circlist, outlist, aux
+    
+###begin BiRB tools###
+def stabilizer_to_all_zs(stabilizer):
+    n = len(stabilizer)
+    
+    symp_reps = _symp.compute_internal_gate_symplectic_representations()
+    
+    s_inv_p, p_inv_p = _symp.inverse_clifford(symp_reps['P'][0],symp_reps['P'][1])
+    s_h, p_h = symp_reps['H']
+    s_y, p_y = symp_reps['C1']
+    
+    stab_layer = []
+    c_str = [[]]
+    
+    for i in range(n):
+        if stabilizer[i] == 'Y':
+            stab_layer.append((s_y, p_y))
+            c_str[0].append(('Gc1','Q{}'.format(i)))
+        elif stabilizer[i] == 'X':
+            stab_layer.append((s_h, p_h))
+            c_str[0].append(('Gc12','Q{}'.format(i)))
+        elif stabilizer[i] == 'I':
+            rand_clifford = str(_np.random.choice(_np.arange(24)))
+            s_rand, p_rand = symp_reps['C'+rand_clifford]
+            stab_layer.append((s_rand, p_rand))
+            c_str[0].append(('Gc'+rand_clifford,'Q{}'.format(i)))
+        else:
+            s_rand, p_rand = symp_reps['C0']
+            stab_layer.append((s_rand, p_rand))
+            c_str[0].append(('Gc0', 'Q{}'.format(i)))
+            
+    s_layer, p_layer = _symp.symplectic_kronecker(stab_layer)
+    stab_circuit = _cir.Circuit(c_str).parallelize()
+    
+    return s_layer, p_layer, stab_circuit
+
+def symplectic_to_pauli(s,p):
+    # Takes in the symplectic representation of a Pauli (ie a 2n bitstring in the Hostens notation) and converts it into a list of characters
+    # representing the corresponding stabilizer.
+    #     - s: Length 2n bitstring.
+    #     - p: The "global" phase.
+    # Returns: A list of characters ('I','Y','Z','X') representing the stabilizer that corresponds to s.
+    
+    n = int(len(s)/2)
+    pauli = []
+    for i in range(n):
+        x_pow = s[i]
+        z_pow = s[n+i]
+        if x_pow != 0 and z_pow != 0: # Have XZ in the i-th slot, ie product is a Y
+            #print('need to undo a Y, apply HP^(-1)')
+            pauli.append('Y')
+        elif x_pow != 0 and z_pow == 0: # Have X in the i-th slot, ie product is an X
+            #print('need to undo an X, so apply inverse Hadamard, ie a Hadamard')
+            pauli.append('X')
+        elif x_pow == 0 and z_pow != 0: # Have Z or I in the i-th slot, so nothing needs to be done
+            #print('need to undo a Z or I, ie leave it be')
+            pauli.append('Z')
+        else:
+            pauli.append('I')
+            
+    return pauli
+
+def sample_random_pauli(n,pspec = None, absolute_compilation = None, qubit_labels = None, circuit = False, include_identity = False):
+    # Samples a random Pauli along with a +-1 phase. Returns the Pauli as a list or as a circuit depending 
+    # upon the value of "circuit"
+    #     - n: Number of qubits
+    #     - pspec: Processor spec
+    #     - absolute_compilation: compilation rules 
+    #     - qubit_labels:
+    #     - circuit: Boolean that determines if a list of single-qubit Paulis or a compiled circuit is returned.
+    
+    if circuit is True:
+        if qubit_labels is not None: qubits = qubit_labels[:]  # copy this list
+        else: qubits = pspec.qubit_labels[:]
+    
+    pauli_list = ['I','X','Y','Z']
+    
+    if include_identity is False:
+        while True: 
+            rand_ints = _np.random.randint(0,4,n)
+            if sum(rand_ints) != 0: # make sure we don't get all identities
+                break
+    else:
+        rand_ints = _np.random.randint(0, 4, n)
+            
+    pauli = [pauli_list[i] for i in rand_ints]
+    if set(pauli) != set('I'): sign = _np.random.choice([-1,1])
+    else: sign = 1
+    
+    if circuit is False:
+        return pauli, sign
+    else:
+        pauli_layer_std_lbls = [_lbl.Label(pauli_list[rand_ints[q]], (qubits[q],)) for q in range(n)]
+        # Converts the layer to a circuit, and changes to the native model.
+        pauli_circuit = _cir.Circuit(layer_labels=pauli_layer_std_lbls, line_labels=qubits).parallelize()
+        pauli_circuit = pauli_circuit.copy(editable=True)
+        pauli_circuit.change_gate_library(absolute_compilation)
+        if pauli_circuit.depth == 0:
+            pauli_circuit.insert_layer_inplace([_lbl.Label(())], 0)
+
+        pauli_circuit.done_editing()
+    return pauli, sign, pauli_circuit
+
+      
+def select_neg_evecs(pauli, sign):
+    # Selects the entries in an n-qubit that will be turned be given a -1 1Q eigenstates
+    #     - pauli: The n-qubit Pauli
+    #     - sign: Whether you want a -1 or +1 eigenvector
+    # Returns: A bitstring whose 0/1 entries specify if you have a +1 or -1 1Q eigenstate
+    n = len(pauli)
+    identity_bitstring = [0 if i == 'I' else 1 for i in pauli]
+    nonzero_indices = _np.nonzero(identity_bitstring)[0]
+    num_nid = len(nonzero_indices)
+    if num_nid % 2 == 0:
+        if sign == 1:
+            choices = _np.arange(start = 0, stop = num_nid+1, step = 2)
+        else:
+            choices = _np.arange(start = 1, stop = num_nid, step = 2)
+    else:
+        if sign == 1:
+            choices = _np.arange(start = 0, stop = num_nid, step = 2)
+        else:
+            choices = _np.arange(start = 1, stop = num_nid+1, step = 2)
+    num_neg_evecs = _np.random.choice(choices)
+    assert((-1)**num_neg_evecs == sign)
+    
+    neg_evecs = _np.random.choice(nonzero_indices, num_neg_evecs, replace = False)
+    
+    bit_evecs = _np.zeros(n)
+    bit_evecs[neg_evecs] = 1
+    assert('I' not in _np.array(pauli)[nonzero_indices])
+    
+    return bit_evecs
+
+def compose_initial_cliffords(prep_circuit):
+    composition_rules = {'Gc0': 'Gc3',
+                         'Gc2': 'Gc5',
+                         'Gc12': 'Gc15'} #supposed to give Gc# * X
+    
+    sign_layer = prep_circuit[0]
+    circ_layer = prep_circuit[1]
+    composed_layer = []
+    
+    for i in range(len(sign_layer)):
+        sign_gate = sign_layer[i]
+        circ_gate = circ_layer[i]
+        new_gate = circ_gate
+        if sign_gate == 'Gc3': # we know that what follows must prep a X, Y, or Z stablizer
+            new_gate = composition_rules[circ_gate]
+        composed_layer.append(new_gate)
+    return composed_layer
+
+def sample_stabilizer(pauli, sign):
+    # Samples a random stabilizer of a Pauli, s = s_1 \otimes ... \otimes s_n. For each s_i,
+    # we perform the following gates:
+    #     - s_i = X: H
+    #     - s_i = Y: PH
+    #     - s_i = Z: I
+    #     - s_i = I: A random 1Q Clifford
+    # Also creates the circuit layer that is needed to prepare
+    # the stabilizer state. 
+    #     - pauli: a list of 1Q paulis whose tensor product gives the n-qubit Pauli
+    # Returns: The symplectic representation of the stabilizer state, symplectic representation of the 
+    #          preparation circuit, and a pygsti circuit representation of the prep circuit
+    n = len(pauli)
+    neg_evecs = select_neg_evecs(pauli, sign)
+    assert((-1)**sum(neg_evecs) == sign)
+    zvals = [0 if neg_evecs[i] == 0 else -1 for i in range(n)]
+    
+    #zvals = _np.random.choice([0,-1], n) 
+    #print(zvals)
+    
+    # init_stab, init_phase = _symp.prep_stabilizer_state(n, zvals)
+    init_stab, init_phase = _symp.prep_stabilizer_state(n)
+    
+    symp_reps = _symp.compute_internal_gate_symplectic_representations()
+    
+    layer_dict = {'X': symp_reps['H'],
+                'Y': tuple(_symp.compose_cliffords(symp_reps['H'][0] 
+                                                   ,symp_reps['H'][1]
+                                                   ,symp_reps['P'][0]
+                                                   ,symp_reps['P'][1])), 
+                'Z': symp_reps['I']}
+    circ_dict = {'X': 'Gc12',
+                 'Y': 'Gc2',
+                 'Z': 'Gc0'}
+    
+    x_layer = [symp_reps['I'] if zvals[i] == 0 else symp_reps['X'] for i in range(len(zvals))]
+    
+    circ_layer = [circ_dict[i] if i in circ_dict.keys() else 'Gc'+str(_np.random.randint(24)) for i in pauli]
+    #init_layer = [layer_dict[pauli[i]] if pauli[i] in layer_dict.keys() else symp_reps[circ_layer[i].replace('Gc','C')] for i in range(len(pauli))]
+    
+    init_layer = [symp_reps[circ_layer[i].replace('Gc', 'C')] for i in range(len(pauli))]
+                  
+    x_layer_rep, x_layer_phase = _symp.symplectic_kronecker(x_layer)
+
+    layer_rep, layer_phase = _symp.symplectic_kronecker(init_layer)
+    
+    #stab_state, stab_phase =  _symp.apply_clifford_to_stabilizer_state(layer_rep, layer_phase,
+    #                                                                  stab_state, stab_phase) 
+    
+    s_prep, p_prep = _symp.compose_cliffords(x_layer_rep,x_layer_phase, layer_rep, layer_phase)
+    
+    stab_state, stab_phase = _symp.apply_clifford_to_stabilizer_state(s_prep, p_prep, init_stab, init_phase)
+    
+    sign_layer = ['Gc0' if zvals[i] == 0 else 'Gc3' for i in range(len(zvals))]
+    
+    layer = [sign_layer, circ_layer]
+    
+    # sign = (-1)**(-1*_np.sum(zvals))
+
+    #return stab_state, stab_phase, layer_rep, layer_phase, circuit_rep
+    return stab_state, stab_phase, s_prep, p_prep, layer
+
+def measure(s_state, p_state):
+    num_qubits = len(p_state) // 2
+    outcome = []
+    for i in range(num_qubits):
+        p0, p1, ss0, ss1, sp0, sp1 = _symp.pauli_z_measurement(s_state, p_state, i)
+        # could cache these results in a FUTURE _stabilizer_measurement_probs function?
+        if p0 != 0:
+            outcome.append(0)
+            s_state, p_state = ss0, sp0 % 4
+        else:
+            outcome.append(1)
+            s_state, p_state = ss1, sp1 % 4
+    return outcome
+
+def determine_sign(s_state, p_state, measurement):
+    an_outcome = measure(s_state, p_state)
+    sign = [-1 if bit == 1 and pauli == 'Z' else 1 for bit, pauli in zip(an_outcome, measurement)]
+    return _np.prod(sign) 
+
+
+def create_binary_rb_circuit(pspec, clifford_compilations, length, qubit_labels=None, sampler='Qelimination',
+                             samplerargs=[], addlocal=False, lsargs=[],
+                             citerations=20, compilerargs=[], partitioned=False, seed=None):
+    
+    if qubit_labels is not None: n = len(qubit_labels)
+    else: 
+        n = pspec.num_qubits
+        qubit_labels = pspec.qubit_labels
+
+    rand_state = _np.random.RandomState(seed)  # Ok if seed is None
+    #s_start, p_start = _symp.prep_stabilizer_state(n)
+    
+
+    rand_pauli, rand_sign, pauli_circuit = sample_random_pauli(n = n, pspec = pspec, 
+                                                                   absolute_compilation = clifford_compilations['absolute'],
+                                                                   circuit = True, include_identity = False)
+    
+    s_inputstate, p_inputstate, s_init_layer, p_init_layer, prep_circuit = sample_stabilizer(rand_pauli, rand_sign)
+    prep_circuit = compose_initial_cliffords(prep_circuit)
+    s_pc, p_pc = _symp.symplectic_rep_of_clifford_circuit(pauli_circuit, pspec = pspec)
+    
+    # build the initial layer of the blown up circuit
+    initial_circuit = _cir.Circuit([[(prep_circuit[i], qubit_labels[i]) for i in range(len(qubit_labels))]])
+    full_circuit = initial_circuit.copy(editable = True)
+
+    # Sample a random circuit of "native gates".
+    if sampler == 'Qelimination' or sampler == 'edgegrab':
+        circuit = create_random_circuit(pspec=pspec, length=length, qubit_labels=qubit_labels, sampler=sampler,
+                                    samplerargs=samplerargs, addlocal=addlocal, lsargs=lsargs, rand_state=rand_state)
+    elif sampler == create_random_quintuple_layered_circuit:
+        circuit = sampler(pspec, qubit_labels, length, *samplerargs, rand_state = rand_state)
+    else:
+        raise ValueError(f'{sampler} is not supported')
+        
+    # find the symplectic matrix / phase vector this "native gates" circuit implements.
+    s_rc, p_rc = _symp.symplectic_rep_of_clifford_circuit(circuit, pspec=pspec)
+    
+    s_composite, p_composite = _symp.compose_cliffords(s1 = s_init_layer, p1 = p_init_layer, s2 = s_rc, p2 = p_rc)
+
+    # Apply the random circuit to the initial state (either the all 0s or a random stabilizer state)
+    
+    full_circuit.append_circuit_inplace(circuit)
+    
+    s_outputstate, p_outputstate = _symp.apply_clifford_to_stabilizer_state(s_rc, p_rc, 
+                                                                            s_inputstate, p_inputstate)
+    
+
+    # Figure out what stabilizer of s_outputstate, rand_pauli was mapped too
+    s_rc_inv, p_rc_inv = _symp.inverse_clifford(s_rc, p_rc) # U^(-1)
+    s_new_pauli, p_new_pauli = _symp.compose_cliffords(s_rc_inv, p_rc_inv, s_pc, p_pc) # PU^(-1)
+    s_new_pauli, p_new_pauli = _symp.compose_cliffords(s_new_pauli, p_new_pauli, s_rc, p_rc) # UPaU^(-1)
+        
+    pauli_vector = p_new_pauli
+    pauli = [i[0] for i in _symp.find_pauli_layer(pauli_vector, [j for j in range(n)])]
+    measurement, phase = ['I' if i == 'I' else 'Z' for i in pauli], None #not needed
+        
+    # Turn the stabilizer into an all Z and I stabilizer. Append this to the circuit.
+    
+    s_stab, p_stab, stab_circuit = stabilizer_to_all_zs(pauli)
+    
+    full_circuit.append_circuit_inplace(stab_circuit)
+    
+    s_inv, p_inv = _symp.inverse_clifford(s_stab, p_stab)
+    s_cc, p_cc = _symp.compose_cliffords(s_inv, p_inv, s_composite, p_composite)
+    s_cc, p_cc = _symp.compose_cliffords(s_composite, p_composite, s_stab, p_stab) # MUPaU^(-1)M^(-1)
+    
+    meas = [i[0] for i in _symp.find_pauli_layer(p_cc, [j for j in range(n)])] # not needed
+     
+    s_outputstate, p_outputstate = _symp.apply_clifford_to_stabilizer_state(s_stab, p_stab, s_outputstate, p_outputstate)
+
+    full_circuit.done_editing()
+    sign = determine_sign(s_outputstate, p_outputstate, measurement)
+    
+    if not partitioned: outcircuit = full_circuit
+    else: outcircuit = [initial_circuit, circuit, stab_circuit]
+        
+    return outcircuit, measurement, sign
+
