+++ pygsti/protocols/__init__.py
@@ -19,3 +19,5 @@ from .rpe import *
 from .stability import *
 from .vb import *
 from .vbdataframe import *
+from .birb import *
+from .birbmcm import *
