+++ pygsti/protocols/rb.py
@@ -844,7 +844,6 @@ class MirrorRBDesign(_vb.BenchmarkingDesign):
         -------
         MirrorRBDesign
         """
-        mapped_circuits_and_idealouts_by_depth = self._mapped_circuits_and_idealouts_by_depth(mapper)
         mapped_qubit_labels = self._mapped_qubit_labels(mapper)
         return DirectRBDesign.from_existing_circuits(mapped_circuits_and_idealouts_by_depth,
                                                      mapped_qubit_labels,
@@ -852,6 +851,86 @@ class MirrorRBDesign(_vb.BenchmarkingDesign):
                                                      self.samplerargs, self.localclifford,
                                                      self.paulirandomize, self.descriptor,
                                                      add_default_protocol=False)
+
+
+
+class BinaryRBDesign(_vb.BenchmarkingDesign):
+    def __init__(self, pspec, clifford_compilations, depths, circuits_per_depth, qubit_labels=None,
+                 sampler='edgegrab', samplerargs=[0.25, ],
+                 addlocal=False, lsargs=(),
+                 citerations=20, compilerargs=(), partitioned=False, descriptor='A DRB experiment',
+                 add_default_protocol=False, seed=None, verbosity=1, num_processes=1):
+
+        if qubit_labels is None: qubit_labels = tuple(pspec.qubit_labels)
+        circuit_lists = []
+        measurements = []
+        signs = []
+
+        if seed is None:
+            self.seed = _np.random.randint(1, 1e6)  # Pick a random seed
+        else:
+            self.seed = seed
+
+        for lnum, l in enumerate(depths):
+            lseed = self.seed + lnum * circuits_per_depth
+            if verbosity > 0:
+                print('- Sampling {} circuits at DRB length {} ({} of {} depths) with seed {}'.format(
+                    circuits_per_depth, l, lnum + 1, len(depths), lseed))
+
+            args_list = [(pspec, clifford_compilations, l)] * circuits_per_depth
+            kwargs_list = [dict(qubit_labels=qubit_labels, sampler=sampler, samplerargs=samplerargs,
+                                addlocal=addlocal, lsargs=lsargs,
+                                citerations=citerations, compilerargs=compilerargs,
+                                partitioned=partitioned,
+                                seed=lseed + i) for i in range(circuits_per_depth)]
+            #results = [_rc.create_direct_rb_circuit(*(args_list[0]), **(kwargs_list[0]))]  # num_processes == 1 case
+            results = _tools.mptools.starmap_with_kwargs(_rc.create_binary_rb_circuit, circuits_per_depth,
+                                                         num_processes, args_list, kwargs_list)
+
+            circuits_at_depth = []
+            measurements_at_depth = []
+            signs_at_depth = []
+            for c, meas, sign in results:
+                circuits_at_depth.append(c)
+                measurements_at_depth.append(meas)
+                signs_at_depth.append(sign)
+
+            circuit_lists.append(circuits_at_depth)
+            measurements.append(measurements_at_depth)
+            signs.append(signs_at_depth)
+
+        self._init_foundation(depths, circuit_lists, measurements, signs, circuits_per_depth, qubit_labels,
+                              sampler, samplerargs, addlocal, lsargs, citerations, compilerargs, partitioned, descriptor,
+                              add_default_protocol)
+
+    def _init_foundation(self, depths, circuit_lists, measurements, signs, circuits_per_depth, qubit_labels,
+                         sampler, samplerargs, addlocal, lsargs, citerations, compilerargs, partitioned, descriptor,
+                         add_default_protocol):
+        super().__init__(depths, circuit_lists, signs, qubit_labels, remove_duplicates=False)
+        self.measurements = measurements
+        self.signs = signs
+        self.circuits_per_depth = circuits_per_depth
+        self.citerations = citerations
+        self.compilerargs = compilerargs
+        self.descriptor = descriptor
+        if isinstance(sampler, str):
+            self.sampler = sampler
+        else:
+            self.sampler = 'function'
+        self.samplerargs = samplerargs
+        self.addlocal = addlocal
+        self.lsargs = lsargs
+        self.partitioned = partitioned
+
+        if add_default_protocol:
+            if randomizeout:
+                defaultfit = 'A-fixed'
+            else:
+                defaultfit = 'full'
+            self.add_default_protocol(RB(name='RB', defaultfit=defaultfit))
+            
+        self.auxfile_types['signs'] = 'json' # Makes sure that signs and measurements are saved seperately
+        self.auxfile_types['measurements'] = 'json'
 
 
 class RandomizedBenchmarking(_vb.SummaryStatistics):
@@ -872,6 +951,8 @@ class RandomizedBenchmarking(_vb.SummaryStatistics):
         S = sum_{k = 0}^n (-1/2)^k h_k where h_k is the frequency at which the output bitstring is
         a Hamming distance of k from the target bitstring, and n is the number of qubits.
         This datatype is used in Mirror RB, but can also be used in Clifford and Direct RB.
+        
+        #added in "energies" as a datatype for DRB without Inversion
 
     defaultfit: 'A-fixed' or 'full'
         The summary data is fit to A + Bp^m with A fixed and with A as a fit parameter.
@@ -921,6 +1002,8 @@ class RandomizedBenchmarking(_vb.SummaryStatistics):
             S = sum_{k = 0}^n (-1/2)^k h_k where h_k is the frequency at which the output bitstring is
             a Hamming distance of k from the target bitstring, and n is the number of qubits.
             This datatype is used in Mirror RB, but can also be used in Clifford and Direct RB.
+            
+            # Added in "energies" to work with Direct RB without Inversion
 
         defaultfit: 'A-fixed' or 'full'
             The summary data is fit to A + Bp^m with A fixed and with A as a fit parameter.
@@ -957,7 +1040,7 @@ class RandomizedBenchmarking(_vb.SummaryStatistics):
         super().__init__(name)
 
         assert(datatype in self.summary_statistics), "Unknown data type: %s!" % str(datatype)
-        assert(datatype in ('success_probabilities', 'adjusted_success_probabilities')), \
+        assert(datatype in ('success_probabilities', 'adjusted_success_probabilities', 'energies')), \
             "Data type '%s' must be 'success_probabilities' or 'adjusted_success_probabilities'!" % str(datatype)
 
         self.seed = seed
@@ -968,6 +1051,10 @@ class RandomizedBenchmarking(_vb.SummaryStatistics):
         self.datatype = datatype
         self.defaultfit = defaultfit
         self.square_mean_root = square_mean_root
+        if self.datatype == 'energies':
+            self.energies = True
+        else:
+            self.energies = False
 
     def run(self, data, memlimit=None, comm=None):
         """
@@ -992,7 +1079,7 @@ class RandomizedBenchmarking(_vb.SummaryStatistics):
         design = data.edesign
 
         if self.datatype not in data.cache:
-            summary_data_dict = self._compute_summary_statistics(data)
+            summary_data_dict = self._compute_summary_statistics(data, energy = self.energies)
             data.cache.update(summary_data_dict)
         src_data = data.cache[self.datatype]
         data_per_depth = src_data
@@ -1010,6 +1097,8 @@ class RandomizedBenchmarking(_vb.SummaryStatistics):
                 asymptote = 1 / 2**nqubits
             elif self.datatype == 'adjusted_success_probabilities':
                 asymptote = 1 / 4**nqubits
+            elif self.datatype == 'energies':
+                asymptote = 0
             else:
                 raise ValueError("No 'std' asymptote for %s datatype!" % self.asymptote)
 
@@ -1023,10 +1112,10 @@ class RandomizedBenchmarking(_vb.SummaryStatistics):
                     adj_sps.append(_np.nanmean(_np.sqrt(percircuitdata))**2)
                     #print(adj_sps)
                 else:
-                    adj_sps.append(_np.nanmean(percircuitdata))  # average [adjusted] success probabilities
+                    adj_sps.append(_np.nanmean(percircuitdata))  # average [adjusted] success probabilities or energies
 
             #print(adj_sps)
-
+            # Don't think this needs changed
             full_fit_results, fixed_asym_fit_results = _rbfit.std_least_squares_fit(
                 depths, adj_sps, nqubits, seed=self.seed, asymptote=asymptote,
                 ftype='full+FA', rtype=self.rtype)
@@ -1034,6 +1123,7 @@ class RandomizedBenchmarking(_vb.SummaryStatistics):
             return full_fit_results, fixed_asym_fit_results
 
         #do RB fit on actual data
+        # Think this works just fine
         ff_results, faf_results = _get_rb_fits(data_per_depth)
 
         if self.bootstrap_samples > 0:
@@ -1078,7 +1168,7 @@ class RandomizedBenchmarking(_vb.SummaryStatistics):
             bootstraps_faf = None
             std_faf = None
             failrate_faf = None
-
+        # we are here
         fits = _tools.NamedDict('FitType', 'category')
         fits['full'] = _rbfit.FitResults(
             'LS', ff_results['seed'], self.rtype, ff_results['success'], ff_results['estimates'],
@@ -1203,9 +1293,11 @@ class RandomizedBenchmarkingResults(_proto.ProtocolResults):
             a = self.fits[fitkey].estimates['a']
             b = self.fits[fitkey].estimates['b']
             p = self.fits[fitkey].estimates['p']
+            #_plt.plot(lengths, a + b * p**lengths,
+            #          label='Fit, r = {:.2} +/- {:.1}'.format(self.fits[fitkey].estimates['r'],
+            #                                                  self.fits[fitkey].stds['r']))
             _plt.plot(lengths, a + b * p**lengths,
-                      label='Fit, r = {:.2} +/- {:.1}'.format(self.fits[fitkey].estimates['r'],
-                                                              self.fits[fitkey].stds['r']))
+                      label='Fit, r = {:.2}'.format(self.fits[fitkey].estimates['r']))
 
         if success_probabilities:
             all_success_probs_by_depth = [data_per_depth[depth] for depth in self.depths]
@@ -1224,6 +1316,7 @@ class RandomizedBenchmarkingResults(_proto.ProtocolResults):
         else: _plt.show()
 
         return
+    
 
     def copy(self):
         """
